+++ src/node.rs
@@ -38,6 +38,7 @@ pub const REMOTE_NODES: [(&str, &str, &str, &str); 16] = [
 	("monero.10z.com.ar",           "AR - Buenos Aires F.D.",         "18089", "18084"),
 	("monero1.heitechsoft.com",     "CA - Ontario",                   "18081", "18084"),
 	("node.monerodevs.org",         "CA - Quebec",                    "18089", "18084"),
+	("node.cryptocano.de",          "DE - Lower Saxony",              "18089", "18083"),
 	("p2pmd.xmrvsbeast.com",        "DE - Hesse",                     "18081", "18083"),
 	("node2.monerodevs.org",        "FR - Occitanie",                 "18089", "18084"),
 	("p2pool.uk",                   "GB - England",                   "18089", "18084"),