+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 	- Removed `de.poiuty.com`
 	- Removed `m1.poiuty.com`
 	- Removed `reynald.ro`
+	- Removed `monero.homeqloud.com`
 
 ## Fixes
 *