+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Updates
 * [Remote Node](https://github.com/hinto-janai/gupax#remote-monero-nodes) changes:
 	- Added `sf.xmr.support`
+	- Added `node.cryptocano.de`
 	- Removed `xmr.aa78i2efsewr0neeknk.xyz`
 	- Removed `node.yeetin.me`
 	- Removed `monero2.10z.com.ar`