+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * [Remote Node](https://github.com/hinto-janai/gupax#remote-monero-nodes) changes:
 	- Added `sf.xmr.support`
 	- Added `node.cryptocano.de`
+	- Added `bunkernet.ddns.net`
 	- Removed `xmr.aa78i2efsewr0neeknk.xyz`
 	- Removed `node.yeetin.me`
 	- Removed `monero2.10z.com.ar`