+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 	- Added `node.cryptocano.de`
 	- Added `bunkernet.ddns.net`
 	- Added `oracle.netrix.cc`
+	- Added `fbx.tranbert.com`
 	- Removed `xmr.aa78i2efsewr0neeknk.xyz`
 	- Removed `node.yeetin.me`
 	- Removed `monero2.10z.com.ar`